requirements.txt
bs4
attrs
icalendar
httpx
